user_data: mark sizes as capped after a memcpy()
This code is to reduce duplicate warnings.  If you have a
memcpy(dst, src, user_size) then only warn about the first memcpy().

I possibly should get the max size of the dst buffer and set the state to
that, but this is a more conservative approach.  I want to get a release
out soon so I'm trying to avoid breakage.

Also use the select_return_param_key() for this because it's cleaner.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>