db: do not split partially impossible returns by positive/negative
Most code in the kernel returns zero or negative error codes.  The problem
is that say you have a function which returns an empty state.  The
is_kernel_success_fail() function refuses to parse that.  So then it
gets parsed by split_positive_from_negative() which is the worst way to
split the code.

Check for empty states and don't do a pos/neg split for possibly empty
states.  I think there are no implications for that anyway so it just
duplicates the same states three times (waste of space) for no reason.

(Really splitting never makes sense unless there are implications.  For
a future patch).

Now it's handled by split possible/impossible which is also not really
ideal but it's an improvement.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>