points_to_user_data: differentiate better from set vs transferred data
The points_to_user_data module differentiates between data which was
copied from the user in the function itself vs data which is just
passed in and then returned or whatever.  But the set_points_to_user_data()
function promoted it all to set data.

The impact of that was that somehow Smatch thought users were controlling
things inside spin_locks.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>