user_data: track assigned states more accurately
There are two reasons that a function could return user data.  1) It gets
data from the user and returns that (USER_DATA_SET).  2)  We pass in user
data and it does "return user_data;" (USER_DATA).  The code here is for
when we get user data from one parameter and copy it to the other.

	dest->foo = src->foo;

Where "src->foo" might be user data.  The problem is that it's using
!param_was_set_var_sym(sm->name, sm->sym) to determine if the variable
was set locally.  But imagine that "src->foo" is user data still and we
do:

	src->foo = 1;

The param_was_set_var_sym() function returns true, and it records that
we set that parameter to user data even though we didn't.  So track
assignments separately and more accurately.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>