buf_comparison: don't use assigned variable so much
The bug here is that we have:

	len = min_t();
	...
	buf = kmalloc(len, GFP_KERNEL);

Originally it would record that "buf" is size min_t() which is accurate but
not as useful as saying that it's "len".  We do still want to handle
"len = nr * sizeof();" so we track those assignments still.

A more ideal solution would be to record both relationships.  #IDEA

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>