struct_assign: don't ignore unknown types
A different title for this patch would be "ignore unknown types *HARDER*".

The problem is that we have code like:

	buffer = (unsigned char *)__get_free_page(GFP_KERNEL);

It should create a fake assignment something like:

	*buffer = *(unsigned char *)__get_free_page(GFP_KERNEL);

But because __get_free_page() returns an unsigned long instead of a pointer
then this code was doing "return expr;" instead of
"return deref_expression(expr);".  So that meant it created a fake
assignment like:

	*buffer = __get_free_page(GFP_KERNEL);

And then Smatch asks "Where is the release function?".  So instead of that
just return NULL.  That means that no fake "*buffer" assignment is created.
Which is probably fine.

I will re-evaluate if this causes new false positives.  I'll probably
return a fake call here or something.  But let's test it and see if that's
required at all.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>