buf_comparison: use comparison data when handling assignments
The real meat of this patch is in match_assign_smaller().  What we have
is something like:

	buf = alloc(len);

	if (len > foo)
		len = foo;

At the end of that code len is still the size of buf (actually equal or
smaller but this code doesn't have a way to represent smaller than buf
size only equal to).  So if we assign a smaller value to len then it's
still the valid.

This information is stored in smatch_comparison.  The problem is we're
handling an assignment in smatch_buf_comparison and smatch_comparison is
handling the assignment as well at the same time.  There is an ordering
issue.  As well the modification hooks are handling the assignment.

So this assignment needs to be handled before the modification hooks.  Then
we need to ignore it when it comes to updating the assignments.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>