buf_comparison: ignore fake parameter assignments
The problem is that this code assumes that every buffer has one "len"
variable.  So then when we have an assignment "foo->len = len;" it assumes
that people will use "foo->len" instead of "len" from then on.  It is a
stupid assumption which needs to be fixed.

But it's especially stupid for fake parameter assignments and that's a one
liner fix.  So do that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>