returns_negative_error_code: new module to track negative error codes
Most of the time, you can kind of guess that "if a function returns int
and one path is zero and the other path is every non-zero value and it's
a kernel function then probably the non-zero side is an error path."

That's generally how humans read code so it's kind of fine but it's also
a little bit guessy.  And it's really simple enough to just track which
functions return error codes.

That's what this module does.  I have several unpublished checks which use
this information and I can already see that it's pretty useful so I'm
publishing this code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>