impossible: use comparison information to set impossible paths
Before if we had something code like:

	x = 3;
	if (x > 3) {
		// <-- this is impossible

But now I want to extend that to comparisons between variables as well,
where we don't know the value of x, but we only know x == foo.

	x = foo;
	if (x > foo) {
		// <-- this is impossible

I had expected that this change would introduce more bugs because it means
we're exercising the smatch_comparison.c code much more.  But so far so
good.

The main thing that this affects is check_uninitialized.c which cares a lot
about impossible paths.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>