buf_comparison: size is basically never zero
The match_assign_smaller() function is supposed to handle code like:

	buf = alloc(size);

	if (count < size)
		size = count;

At the end, size is <= the size of buf, but this code doesn't have a way
to represent <= only == so it says size is the number of bytes.  However
one thing that people do is they say:

	buf = alloc(tmp);
	size = tmp;

	tmp = 0;
	blah blah tmp++ something.

And since 0 is <= the original value of tmp it records that tmp is still
the number of bytes.  Which is nonsense.

It probably would have been a 95% solution to ignore zero assignments, but
I just decided to ignore all fixed size assignments.  Once you move from
variable to fixed size that is handled by smatch_buf_size.c instead of
smatch_buf_comparison.c.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>