inline: comment about creating node of node on variadics
When inlining a variadic function the extra arguments are
added in the declaration list as SYM_NODE but these arguments
can already be SYM_NODEs.

Sparse doesn't support everywhere such nested nodes (they must
be merged) but in this case it's fine as the node will be merged
when evaluated.

Add a comment telling the situation is fine.
Also, move the code to where the variadic arguments are handled
since the fixed one will be anyway directly overwritten.

Note: Sparse doesn't really support inlining of variadic functions
      but is fine when the arguments are not used (and such cases
      occur in the kernel).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>