inline: avoid needless intermediate vars
In inline_function(), we need to iterate over the parameters
and the (effective) arguments. An itermediate variable is used for
each: "name_list" and "arg_list".

These confuse me a lot (especially "name_list", "param_list" would
be much more OK) and are just used once.

So, avoid using an intermediate variable for these.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>