points_to_host_data: differentiate between set pointer and passed in pointers
This copies the code from smatch_points_to_user_data.c.  The issue is that
sometimes if we're returning host data it's because the function reads data
from the host and sometimes it's because the function just returned data
that was passed to it.

The problem is that if we don't differentiate between these two types of
data then eventually more and more data gets marked as host data and it
ends up taking a huge amount of disk space.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>