extra: more fake assignment/mtag fixes
Reading the commit log, ignoring fake assignments or not has kind of been
treated as an optimization, but it's not.  One thing that happens is we
do:

	struct foo *p = &some_struct;

In that case, we don't know the value of some_struct.member because we have
not built the database.  So then it means we're creating a fake assignment
p->a = some_struct.a;  Which is the same thing as "some_struct.a =
some_struct.a;"  It records in the mtag_data that some_struct.a is set to
unknown.

So this kind of assignment needs to be ignored for the flow analysis to
work.  This was detected by the validation/ tests.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>