struct_assignment: handle BUF_CLEAR for base types
One of the things that this is supposed to handle is when we have:

	clear(&my_int);

But the db_buf_cleared() passes a NULL parameter which gets passed down
the line to handle_non_struct_assignments() and makes it a no-op.

The check for "if (right)" doesn't make sense at that point.  Do it
earlier.  And I don't feel like deref_expression(right); make sense either
so just delete that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>