buf_comparison2: new module for linking buffers with their size
The buf_comparison module tracks a bunch of different ways that people use
to specify how large a buffer is.  They can say it has this many elements,
this many bytes, or the last element is x.  But the problem with it is that
it assumes that there is one variable which is the buffer and one variable
which is the limit.  But in real life, people use multiple variables or
they say that x is <= to the limit.

So buf_comparison2 just tracks byte sizes and it uses the smatch_comparison.c
module to track the relationships.  What we do is we create an sm_state
"$size buf vs byte_count" and then if we have a variable which is
< byte_count the comparison module will automatically track that it's
"$size buf".

This kind of duplicates some of the work in smatch_buf_comparison.c but
that's fine.  Eventually I maybe will re-write it all to work this way.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>