flow: clean up how impossible case statements are handled
If you have a bunch of case statements in a row:

	case 4:
	case 7:
	case 1:

And the middle one, case 7 in this example, is impossible, then continue
parsing case 1 instead of skipping to the end.  I think this was sort of
handled already because when we exited out of this loop then the next
statement we parsed would be case 1?  But then why have the loop at all?

More questions than answers really...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>