assigned_expr: clear out states better
The issue here was there was code like:

	ret = ib_copy_from_udata(&ureq, udata,
				 min(udata->inlen, sizeof(ureq)));

And the min() macro translates into screens full of text.  So expr_to_str()
refuses to print it out.  That means that alloc_state_expr() returns NULL.
So then it's marked as handled but really the set_state() is a no-op.

So test if alloc_state_expr() returns NULL and don't set "skip_mod" if
that happens.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>