math: fix handling of binop overflows for user data
If you have "a + b" and they overflow, then get_rl type functions were
returning unknown.  Which is sort of fine for most things.  Implied is
not really true, and absolute will just be the whole range.  But where
it breaks down is for user data.  Instead just say that if it overflows
then it is the whole range.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>