kernel_user_data: make get_user_rl() recursable
Is recursable the right word?  Is it even a word?  The problem is that
var_user_rl() calls get_user_rl() so that means get_user_rl() calls
itself.

I hadn't intended this to be possible, but it seems like people will want
to do it so lets ensure that it works correctly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>