estate: handle capped for empty states better
TLDR; This is basically where I changed .value to .uvalue.

The capped flag means that a variable is capped to an unknown limit.
"if (x > unknown) return -EINVAL;".  What this is supposed to do is say
that if we have two states and one is capped to a very low known limit
and the other is capped with a unknown limit then the result is capped to
an unknown limit.  That kind of makes sense conceptually, I think.

Then the other issue is that if you have an impossible/empty state that's
considered capped.  Which also makes sense conceptually.  If you merge a
capped state with an impossible state the result is capped.

But the upper bound should be unsigned instead of signed.  So use .uvalue
instead of .value.  Also make the check for empty states explicit because
it was less clear what an empty state will return for estate_max().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>