helper: make strip_parens_symbol() strip everything instead
This is a hack.  The function is clearly not meant to strip everything.

But I wrote this hack to fix a bug and I've been running it for probably
most of a year now but I'm just to embarrassed to publish it.  I don't
remember what bug it fixes.  I'm just going to publish it so that my diff
with the released code gets smaller.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>