extra: handle PARAM_SET for buffers better
Quite often for buffers we set one element at a time in a loop:

	for (i = 0; i < len; i++)
		read_one_element(buf + i);

The read_one_element() function sets *$ but that doesn't translate well
into the caller because Smatch doesn't handle buffers well.  Ideally
Smatch would say it sets "*(buf[0 to 9])" or something.  But that's hard.
What we can do instead is say that it sets *buf.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>