user/host_data: assume capped data is less than INT_MAX
What happens here is that we have:

	unsigned int user_data = user_value;
	int signed_data;

	if (user_data > limit)
		return -EINVAL;

	signed_data = user_data;
	__smatch_user_rl(signed_data);

The original user_data is unsigned and it's capped to an unknown value.  It
is probably safe to assume that "limit" is smaller than INT_MAX.  Which
means that "signed_data" can be represented as 0-s32max[capped].

Otherwise it creates false positives about negative user values which are
extremely unlikely.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>