extra: don't over write previous if statements because of dereference data
The issue here looks this:

void frob(struct whatever *p)
{
	if (p)
		p->foo = 1;
}

Some callers pass "p" as non-NULL and some as NULL.  Non-NULL is not the
same as "valid" because that includes error pointers for example.  So
dereferencing "p" is different from just checking for NULL and we want to
change the state of "p" from non-NULL to valid.

The other thing to note is that this function is void and the return
states will be split on the p is NULL vs non-NULL.  So we want to record
the PARAM_LIMIT of "p" because that's pretty central for understanding
the return values.

This code in smatch extra is supposed to say "Let's ignore states that
come from dereferences" because it's not really meaningful to create a
PARAM_LIMIT from dereference.

However, if we already have a state then recording a PARAM_LIMIT is the
correct behavior.

One thing that I wonder about is when the original state is 0-u64max and
the PARAM_LIMIT is 0,4096-ptr_max.  Is that useful information?  Still,
I'm going to try this code and apply fixes on top of it if it causes
issues.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>