check_host_input: add a new pattern
Smatch pattern to facilitate the hardening of the Linux guest kernel
for Confidential Cloud Computing threat model.
In this model the Linux guest kernel cannot trust the values
it obtains using low level IO functions because they can be provided
by a potentially malicious host or VMM. Instead it needs to make
sure the code that handles processing of such values is hardened,
free of memory safety issues and other potential security issues.

This smatch pattern helps to indentify such places.
Currently it covers most of MSR, portIO, MMIO, PCI config space
and cpuid reading primitives.
The full list of covered functions is stored in host_input_funcs array
in the smatch_kernel_host_data.c file.
The output of the pattern can be used to facilitate code audit, as
well as to verify that utilized fuzzing strategy can reach all the
code paths that can take a low-level input from a potentially malicious host.

When ran, the pattern produces two types of findings: errors and warnings.
This is done to help prioritizing the issues for the manual code audit.
However, if time permits, all locations reported by the pattern should be checked.

Signed-off-by: Elena Reshetova <elena.reshetova@intel.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>