mtag: switch to sha1 hashes
These openssl functions are deprecated.  Might as well change to sha1 as
well.  Smatch doesn't use hashes for security so md5 was fine and sha1 is
even *more* fine.  But apparently newer CPUs have operations to speed up
sha1 hashing so it's faster.  My CPU is old and doesn't have the new
operations but in my testing md5sum and sha1 take exactly the same time.

I probably last benchmarked the difference in the 2001/2002 and twenty
years ago md5 was faster than sha1.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>