db: use hashes instead of file names
It turns out that probably 40% of the database is just file names.  On my
system the DB is 40GB so that's a lot of data which can be saved.  It takes
me about 5 hours to create the DB so it can end up taking until 9am to do
a test.

This change makes the DB slightly more complicated to work with but it's
definitely worth it because of all the time it saves.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>