zero_to_err_ptr: silence some false positives
Doing "return PTR_ERR(p);" when "p" is NULL is a perfectly valid thing to
do.  However, although it can be valid, most of the time, it is a bug.
Hence this check.

However, when people use the PTR_ERR(p) as part of a condition:

	if (PTR_ERR(p) == -ENOMEM) {

then most of those uses are valid so don't print a warning about those.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>