buf_size: use file id instead of file name
The size_from_db_type() function was using a string instead of an hash
to get an array size from the DB.  Introduce get_file_id() to solve this
problem.

I moved the get_base_file_id() to smatch_flow.c because it felt like maybe
this is where everything belongs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>