function_hooks: fix an ordering issue in add_function_param_key_hook()
This is for assignment hooks.  Normally on these hooks when we're assigning
something we only care about the success path.  On the failure path we
return NULL and that's not interesting, but on the success path then there
are resources that need to be freed and buffer sizes that need to be
tracked etc.

But there are some functions which always succeed and that's the bug here.

In check_unwind() we want to track old_cred = ovl_override_creds().  We
store the creds and then we need to restore them at the end with
revert_creds(old_cred);

The problem is that add_function_hook() happens before the modification
hook on the "old_cred = " assignment.  The check_unwind.c check uses SSA
states so it sets the state correctly.  Then the SSA module sees the
assignment and clears the state.  That means that we cannot revert the
creds because we have overwritten the original "old_creds" with the
assignment.

I added a check for EXPR_ASSIGNMENT in get_parent_assignment() to prevent
a forever loop parsing "foo = bar = frob();".  I should probably just get
rid of get_parent_assignment() instead...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>