function_hooks: do set_fresh_mtag_returns() later
This code set the mtag return at the first INTERN record from the DB.  It
should set it at the end.

The other thing is that clearing the mtag return should not happen at this
point.  In other words remove these assignments:

	db_info->ret_state = NULL;
	db_info->ret_str = NULL;

One of the effects of this bug is that if you wanted to track an allocated
variable which was also a fresh allocation (which generates an mtag) then
matching on &valid_ptr_min, &valid_ptr_max did not work.  That's because
an mtag is a very specific subset of the valid_ptr and we wanted an exact
match.  But with this fix it works.

Signed-off-by: Dan Carpenter <error27@gmail.com>