direct_return_instead_of_goto: warn about missing unwinds
What this does is it looks for a direct return in the middle of a series
of gotos.  The main "false positive" caused by this check is that some
people use a goto just to print an error message:

out:
	pr_err("%s failed! ret=%d\n", __func__, ret);
	return ret;

It's not really a false positive...  I dislike this style of code.  Direct
returns are better.

Suggested-by: Harshit Mogalapalli <harshit.m.mogalapalli@oracle.com>
Signed-off-by: Dan Carpenter <error27@gmail.com>