kernel: add handling for kref_put()
The kref_put() function is called like:

	kref_put(&job->kref, bsg_teardown_job);

If the reference count drops to zero then call bsg_teardown_job().  This
code tests for if it seems possible for the reference count to be zero.
If so then it creates a fake call to bsg_teardown_job(&job->kref).

The problem is that sometimes the author takes two references to an object
and Smatch does not see it.  So it can lead to some use after free false
positives.  My idea is that we should make a dummy kref_free() function
that just prints a stack trace if we drop to zero unexpectedly.  I will
propose this idea right now.

Signed-off-by: Dan Carpenter <error27@gmail.com>