function_hooks: use the correct expression
This function call uses "expr" where "left" was intended so it does
nothing.

What it is supposed to do is if you have a function which returns a
parameter like:

int frob(int x)
{
	if (x)
		return x;
	return -1;
}

The caller does:

	ret = frob(x);

This returns "x" so we create a fake assignment "ret = x;" but we know that
it is non-zero so we have to apply that filter on top to say that "ret" is
non-zero.  But it was applying the filter to "ret = frob(x);" instead of
"ret" so the filter did nothing.

Signed-off-by: Dan Carpenter <error27@gmail.com>