dma_mapping_error: delete obsolete check
The dma_mapping_error() function used to be type bool where it would return
true to indicate there was an error.  There was a common bug where people
assumed it returned an error code.

In the end HCH killed this bug forever by changing the dma_mapping_error()
function to return an error code.  This this check is no longer required.

Signed-off-by: Dan Carpenter <error27@gmail.com>