scheduling_in_atomic: Add mas_store_gfp() as a special case
Generally the assumption is that when you see a GFP_KERNEL then the kernel
is going to sleep.  The scheduling in atomic() check looks to see if we
have preempt disabled, and then prints a warning.

The mas_store_gfp() function is slightly unusual because it takes GFP_
flag, but you have to hold a certain spinlock, then it drops the lock
before it sleeps.

I felt like it was cleaner to move the checking for this to
scheduling_in_atomic.c, but now that I'm writing this I'm not so certain.
Anyway, it's done now so we're stuck with it.  Forever.

Signed-off-by: Dan Carpenter <error27@gmail.com>