kernel_user_data: ignore nonsense data
The kernel user data does a really kind of stupid thing where it's like
"did this function get data from the user directly or was there any
user data passed in".  The problem with this is that if the caller passes
in *any* user data at all then it marks it as getting user data back.

So that's what I think is happening.  But sometimes, Smatch thinks that
users can control the deep internal kernel data like spinlock data which
no one should be accessing directly.  (They should be using accessor
functions and not modifying the data directly).

Smatch tries to ignore this data because it's a lot of data and not useful
to the callers.  Ignore it if it's marked as user data as well.

This is a kind of ugly hack, but it's easy to write and saves quite a bit
of space in the DB.

Signed-off-by: Dan Carpenter <error27@gmail.com>