buf_comparison: add support for struct_size() array
This code is supposed to parse stuff like:

	buf = malloc(size * elem_size);

Where the size is not fixed so then if we have:

	buf[x] = foo;

So long as x < size it's fine.  The problem is that sometimes we allocate
a struct with a variable length array:

	my_struct = kmalloc(struct_size(type, array, count), GFP_KERNEL);

So then we need to record that "my_struct.array" as having count elements.

Signed-off-by: Dan Carpenter <error27@gmail.com>