function_hooks: do early stuff even earlier
What I want to do here is use this to hard code stuff into the caller_info
table.  So for example handle_ksmbd_work() passes foo->bar user data to
__handle_ksmbd_work().  I want to hook into the caller right before we
enter stuff into the DB and then set the foo->bar state to user data.
Then when we insert the state into the DB, it's user data so it gets
inserted.

Signed-off-by: Dan Carpenter <error27@gmail.com>