points_to_user_data: communicate USER_PTR returns to USER_DATA
The problem is if a function returns user data as a void pointer.  There
are some functions which are hard coded as returning user data pointers
like memdup_user().  But if it's not hardcoded but instead recorded in
the database then that wasn't handled.  It means that when the void
pointer was eventually cast to a struct pointer the struct members were
not marked as user data.

Signed-off-by: Dan Carpenter <error27@gmail.com>