flow: handle struct assignments later
This should have been committed as part of commit 1571fefbd71f
("points_to_user_data: communicate USER_PTR returns to USER_DATA").
The issue is that I want to parse the returns states to know if a function
returns user data.  That way I can fake the struct assignments more
accurately to say if a variable is user data.

Signed-off-by: Dan Carpenter <error27@gmail.com>