flow: revert previous commit
Ugh.  The ordering of this is tricky.  To be honest, none of this is
correct for functions...

The reason that I can't commit this is because the
__fake_struct_member_assignments clear out old user data.  So if there is
a USER_DATA_SET which sets a -1 $->foo to user data then the fake
assignments clear it out.  That's kind of useless.

But really this stuff is totally broken because what about if one return
path sets the struct members to user data but the others don't?  It doesn't
work to just do one assignment.

Signed-off-by: Dan Carpenter <error27@gmail.com>