points_to_user_data: rename set_points_to_user_data() to set_array_user_ptr()
Originally there was just kernel_user_data and not points_to_user_data.
But that had problems and points_to_user_data was created.  But it was
kind of like an addon, and done in an ad-hoc way.  It wasn't totally
rigorous how the two interacted, and I want to fix that.

Most things should probably start at points_to_user_data and then go to
kernel_user_data.  There are quite a few exceptions like kstroul() etc.
But generally the concept is that when things point to user data and also
hold user data then it should be handled by points to user data first and
then get send to kernel user data.

The exception is when we're setting struct members to user data.  If there
is an array then the whole array gets marked as points to user data.

So rename this function because it's for going backwards.

Signed-off-by: Dan Carpenter <error27@gmail.com>