returns_negative_error_code: low positives are not error codes
If you save an error code into a u32 then it still counts as an error
code for our purposes.  So you can have negative error codes which are
positive.  But if the variable is within the 0-INT_MAX range or for 64
bit types if it's in the 0-UINT_MAX range then that's not an error code.

Signed-off-by: Dan Carpenter <error27@gmail.com>