struct_assignment: introduce create_recursive_fake_assignments()
The way that fake struct assignments worked felt very cool because then
you could re-use the assignment handling code instead of special code to
handle memcpy() or whatever.

But there are a few problems.  First some assignments have side effects
like "foo = bar++;"  Second, it can make things slow because now there are
tons of assignments to parse and the user data code only cares about if
the hand side has data but it has to do tons of look ups.  There is also
a problem with the current code because the fake assignments are currently
hooked in once per assignment but for functions assignments should be done
for every return.

On the other hand, this code is quite nice because I have written code to
recurse through struct members several times.  This code is central place
where it can be done once and re-used.

The create_recursive_fake_assignments() function takes a struct and creates
a bunch of fake assignments:

	foo.bar = fake assign();
	foo.baz = fake assign();

You can also pass a data pointer to custom data.

Signed-off-by: Dan Carpenter <error27@gmail.com>