db: fix returned information
The second loop in print_return_struct_info() is supposed to handle the
situation where we do something like this:

void *frob(void *p)
{
	return p;
}

The "p" pointer is both parameter 0 and the return.  The first loop is
supposed to handle that it's a parameter and the second loop is supposed
to handle that it's a return.  But instead the second loop says "it's a
parameter so don't print anything."

Fix this by creating a version of get_param_key_from_var_sym() that is
explicitly for returns.

Signed-off-by: Dan Carpenter <error27@gmail.com>