user_data/ptr: move things from user_data to user_ptr
This commit is only lightly tested.  I probably won't push it right away.

The issue is that when I want to hard code things as user data, should I
do that in add that in user_data or user_ptr?  Sometimes it had to be done
in both.  And then there was an ordering bug where a function returns a
user_ptr but it wasn't hard coded in the returns_pointer_to_user_data()
array then it didn't get translated to user_data correctly.  Some things
copy_from_user() had to be added to both user_data and user_ptr.

So it's a kind of mess and force me to re-think how this stuff works.

Under the new regime, if it can be described as a pointer then it should
be done that way.  For example, kstrtoXX() take a pointer as an argument
so they are handled by user_ptr.  The simple_strtoXX functions return a
value so they are handled by user_data.

Now if we set something to user_ptr then at the end if can then it gets
translated into user_data as well.  This solves and ordering issue handling
data from the DB and now the is_points_to_user_data_fn() code can be
deleted.

I also re-wrote user_ptr to use the param/key API.

Signed-off-by: Dan Carpenter <error27@gmail.com>