user_data: clean up how iov_iter_count is handled
This code is used to handle hard coded return param/key pairs.  At present
that's only iov_iter_count().

The function has code to handle pointer assignments, but pointer
assignments are only supposed to be handled by points_to_kernel_data.c
and not here.

Also in the original code the state was not marked as a new state so that's
a bug.  Fixed now.

Signed-off-by: Dan Carpenter <error27@gmail.com>