user_data: clean up how sscanf is handled
Technically, sscanf() should probably be handled by points_to_user_data.c
but it's bad to get too hung up on rules.  Doing it this way works and
is much more direct.

The good thing about this patch is that it gets rid of the last caller
for tag_as_user_data() so all that code can be deleted.

Signed-off-by: Dan Carpenter <error27@gmail.com>