param_limit: ignore BUF_CLEAR here as well
This change is slightly less clear, but a singly PARAM_CLEAR of a massive
struct can end up creating over 1000 param_limit states so it can cause
quite a bit of slow down.

The param_limit states are used for understanding which return paths are
possible.  So I imagine this might cause some false positives.  But I doubt
it causes enough false positives to notice.

Signed-off-by: Dan Carpenter <error27@gmail.com>